Transfer TimeSpace to dedicated class and cleanup
